requirements.txt
pillow==8.1.0
Flask==1.1.2
numpy==1.19.4
argparse==1.4.0
tensorflow
annoy==1.17.0
google-cloud
google-cloud-storage
google-cloud-pubsub

# For Annoy, logged issue with Spotify, they will need to fix this,
# In the meantime, I've forked, fixed it and installed from
# my fork.
# annoy==1.17.0